+++ tap_ga4/sync.py
@@ -1,25 +1,44 @@
 import hashlib
 import json
 from datetime import datetime, timedelta
-
 import singer
-from singer import Transformer, get_bookmark, metadata, strftime, utils
-
-from google.analytics.data_v1beta.types import DateRange, Dimension, Metric, RunReportRequest
-
-
+from singer import Transformer, get_bookmark, metadata, utils
+from google.analytics.data_v1beta.types import (Metric, Dimension)
 LOGGER = singer.get_logger()
 
+DEFAULT_CONVERSION_WINDOW = 90
+DEFAULT_REQUEST_WINDOW_SIZE = 7
 
-def generate_sdc_record_hash(raw_report, row, start_date, end_date):
+
+def sort_and_shuffle_streams(currently_syncing, selected_streams):
+    """
+    Order selected streams and shuffle if currently_syncing is set.
+    """
+    stream_list = list(selected_streams)
+    sorted_selected_streams = sorted(stream_list, key=lambda x: x.tap_stream_id)
+
+    if not currently_syncing:
+        return sorted_selected_streams
+
+    currently_syncing_idx = None
+    for i, stream in enumerate(sorted_selected_streams):
+        if currently_syncing == stream.tap_stream_id:
+            currently_syncing_idx = i
+            break
+
+    if currently_syncing_idx:
+        return sorted_selected_streams[currently_syncing_idx:] + sorted_selected_streams[:currently_syncing_idx]
+    return sorted_selected_streams
+
+
+def generate_sdc_record_hash(record, dimension_pairs):
     """
     Generates a SHA 256 hash to be used as the primary key for records
     associated with a report. This consists of a UTF-8 encoded JSON list
     containing:
-    - The account_id, web_property_id, profile_id of the associated report
-    - Pairs of ("ga:dimension_name", "dimension_value")
-    - Report start_date value in YYYY-mm-dd format
-    - Report end_date value in YYYY-mm-dd format
+    - The property_id of the associated report
+    - The account_id of the associated report
+    - Pairs of ("dimension_name", "dimension_value")
 
     Start and end date are included to maintain flexibility in the event the
     tap is extended to support wider date ranges.
@@ -28,74 +47,50 @@ def generate_sdc_record_hash(raw_report, row, start_date, end_date):
     REQUIRE a major version bump! As it will invalidate all previous
     primary keys and cause new data to be appended.
     """
-    dimensions_headers = raw_report["reports"][0]["columnHeader"].get("dimensions", [])
-    profile_id = raw_report["profileId"]
-    web_property_id = raw_report["webPropertyId"]
-    account_id = raw_report["accountId"]
-
-    dimensions_pairs = sorted(zip(dimensions_headers, row.get("dimensions", [])), key=lambda x: x[0])
-
-    # NB: Do not change the ordering of this list, it is the source of the PK hash
-    hash_source_data = [account_id,
-                        web_property_id,
-                        profile_id,
-                        dimensions_pairs,
-                        start_date.strftime("%Y-%m-%d"),
-                        end_date.strftime("%Y-%m-%d")]
-
+    hash_source_data = sorted([("property_id", record["property_id"]),
+                               ("account_id", record["account_id"]),
+                               *dimension_pairs])
     hash_source_bytes = json.dumps(hash_source_data).encode('utf-8')
     return hashlib.sha256(hash_source_bytes).hexdigest()
 
 
-def generate_report_dates(start_date, end_date):
-    total_days = (end_date - start_date).days
-    # NB: Add a day to be inclusive of both start and end
-    for day_offset in range(total_days + 1):
-        yield start_date + timedelta(days=day_offset)
-
-def report_to_records(raw_report):
+def generate_report_dates(start_date, end_date, request_window_size):
     """
-    Parse a single report object into Singer records, with added runtime info and PK.
-
-    NOTE: This function assumes one report being run for one date range
-    per request. For optimizations, the structure of the response will
-    change, and this will need to be refactored.
+    Splits date range from start_date to end_date into chunks of request_window_size
+    length and yields each chunk as a tuple (start_date, end_date).
     """
-    # TODO: Handle data sampling keys and values, either in the records or as a separate stream? They look like arrays.
-    # - https://developers.google.com/analytics/devguides/reporting/core/v4/rest/v4/reports/batchGet#ReportData
-    report = raw_report["reports"][0]
-    column_headers = report["columnHeader"]
-    metrics_headers = [mh["name"] for mh in column_headers["metricHeader"]["metricHeaderEntries"]]
-    dimensions_headers = column_headers.get("dimensions", [])
+    range_start = start_date
+    while range_start <= end_date:
+        # NB: Subtract 1 from request_window_size because date range in RunReportRequest is inclusive
+        range_end = range_start + timedelta(days=request_window_size - 1)
+        yield (range_start.strftime("%Y-%m-%d"), min(end_date, range_end).strftime("%Y-%m-%d"))
+        range_start = range_end + timedelta(days=1)
 
-    for row in report.get("data", {}).get("rows", []):
-        record = {}
-        record.update(zip(dimensions_headers, row.get("dimensions", [])))
-        record.update(zip(metrics_headers, row["metrics"][0]["values"]))
 
-        report_date = raw_report["reportDate"]
-        _sdc_record_hash = generate_sdc_record_hash(raw_report, row, report_date, report_date)
-        record["_sdc_record_hash"] = _sdc_record_hash
-
-        report_date_string = report_date.strftime("%Y-%m-%d")
-        record["start_date"] = report_date_string
-        record["end_date"] = report_date_string
-
-        record["account_id"] = raw_report["accountId"]
-        record["web_property_id"] = raw_report["webPropertyId"]
-        record["profile_id"] = raw_report["profileId"]
-
-        yield record
+def row_to_record(report, row, dimension_headers, metric_headers):
+    """
+    Parse a RunReportResponse row into a single Singer record, with added
+    runtime info and PK.
+    """
+    record = {}
+    dimension_pairs = list(zip(dimension_headers, [dimension.value for dimension in row.dimension_values]))
+    record.update(dimension_pairs)
+    record.update(zip(metric_headers, [metric.value for metric in row.metric_values]))
+    record["property_id"] = report["property_id"]
+    record["account_id"] = report["account_id"]
+    record["_sdc_record_hash"] = generate_sdc_record_hash(record, dimension_pairs)
+    return record
 
 
 DATETIME_FORMATS = {
-    "ga:dateHour": '%Y%m%d%H',
-    "ga:dateHourMinute": '%Y%m%d%H%M',
-    "ga:date": '%Y%m%d',
+    "dateHour": '%Y%m%d%H',
+    "dateHourMinute": '%Y%m%d%H%M',
+    "date": "%Y%m%d",
+    "firstSessionDate": "%Y%m%d"
 }
 
 
-def parse_datetime(field_name, value):
+def parse_datetime(field_name, value, fmt=singer.utils.DATETIME_FMT):
     """
     Handle the case where the datetime value is not a valid datetime format.
 
@@ -106,7 +101,7 @@ def parse_datetime(field_name, value):
     """
     is_valid_datetime = True
     try:
-        parsed_datetime = datetime.strptime(value, DATETIME_FORMATS[field_name]).strftime(singer.utils.DATETIME_FMT)
+        parsed_datetime = datetime.strptime(value, DATETIME_FORMATS[field_name]).strftime(fmt)
         return parsed_datetime, is_valid_datetime
     except ValueError:
         is_valid_datetime = False
@@ -121,22 +116,35 @@ def transform_datetimes(report_name, rec):
             rec[field_name], is_valid_datetime = parse_datetime(field_name, value)
             row_limit_reached = row_limit_reached or (not is_valid_datetime and value == "(other)")
     if row_limit_reached:
-        LOGGER.warning(f"Row limit reached for report: {report_name}. See https://support.google.com/analytics/answer/9309767 for more info.")
+        LOGGER.warning("Row limit reached for report: %s. See https://support.google.com/analytics/answer/9309767 for more info.", report_name)
     return rec
 
 
-def get_start_date(config, property_id, state, tap_stream_id):
+def get_report_start_date(config, property_id, state, tap_stream_id):
     """
-    Returns a date bookmark in state for the given stream, or the
-    `start_date` from config, if no bookmark exists.
+    Returns the correct report start date.
+
+    Cases:
+    start_date: bookmark is empty.
+                OR
+                conversion_date is earlier than the start_date AND bookmark is later than start_date.
+
+    bookmark: bookmark earlier than the conversion_date (this could happen if the tap was paused for awhile).
+
+    conversion_date: the conversion_date is after the start_date AND earlier than the bookmark.
     """
-    start = get_bookmark(state,
-                         tap_stream_id,
-                         property_id,
-                         default={}).get('last_report_date',
-                                         config['start_date'])
-    is_historical_sync = start == config['start_date']
-    return is_historical_sync, utils.strptime_to_utc(start)
+    bookmark = get_bookmark(state,
+                            tap_stream_id,
+                            property_id,
+                            default={}).get('last_report_date')
+    start_date = utils.strptime_to_utc(config['start_date'])
+    if not bookmark:
+        return start_date
+    else:
+        bookmark = utils.strptime_to_utc(bookmark)
+        conversion_window = int(config.get("conversion_window", DEFAULT_CONVERSION_WINDOW))
+        conversion_day = utils.now().replace(hour=0, minute=0, second=0, microsecond=0) - timedelta(days=conversion_window)
+    return min(bookmark, max(start_date, conversion_day))
 
 
 def get_end_date(config):
@@ -151,74 +159,47 @@ def get_end_date(config):
     return utils.now().replace(hour=0, minute=0, second=0, microsecond=0)
 
 
-def get_report(client, property_id, report_date, dimensions, metrics):
-    report_date_string = report_date.strftime("%Y-%m-%d")
-    request = RunReportRequest(
-        property=f'properties/{property_id}',
-        dimensions=dimensions,
-        metrics=metrics,
-        date_ranges=[DateRange(start_date=report_date_string, end_date=report_date_string)]
-    )
-
-    return client.run_report(request)
-
-
-def sync_report(client, schema, report, start_date, end_date, state, is_historical_sync):
+def sync_report(client, schema, report, start_date, end_date, request_window_size, state):
     """
-    Run a sync, beginning from either the start_date or bookmarked date,
-    requesting a report per day, until the last full day of data. (e.g.,
-    "Yesterday")
+    Run a sync, beginning from either the start_date, bookmarked date, or
+    (now - CONVERSION_WINDOW) requesting a report per day.
 
     report = {"name": stream.tap_stream_id,
               "property_id": property_id,
+              "account_id": account_id,
               "metrics": metrics,
               "dimensions": dimensions}
     """
     LOGGER.info("Syncing %s for property_id %s", report['name'], report['property_id'])
 
-    all_data_golden = True
-    # TODO: Is it better to query by multiple days if `ga:date` is present?
-    # - If so, we can optimize the calls here to generate date ranges and reduce request volume
-    for report_date in generate_report_dates(start_date, end_date):
-        for raw_report_response in get_report(client, report["property_id"], report_date, report["dimensions"], report["metrics"]):
+    for range_start_date, range_end_date in generate_report_dates(start_date, end_date, request_window_size):
+        for response in client.get_report(report, range_start_date, range_end_date):
+            dimension_headers = [dimension.name for dimension in response.dimension_headers]
+            metric_headers = [metric.name for metric in response.metric_headers]
             with singer.metrics.record_counter(report['name']) as counter:
-                time_extracted = singer.utils.now()
                 with Transformer() as transformer:
-                    for rec in report_to_records(raw_report_response):
+                    for row in response.rows:
+                        time_extracted = singer.utils.now()
+                        rec = row_to_record(report, row, dimension_headers, metric_headers)
                         singer.write_record(report["name"],
                                             transformer.transform(
                                                 transform_datetimes(report["name"], rec),
                                                 schema),
                                             time_extracted=time_extracted)
                         counter.increment()
+        singer.write_bookmark(state,
+                              report["id"],
+                              report["property_id"],
+                              {"last_report_date": range_end_date})
+        singer.write_state(state)
+    LOGGER.info("Done syncing %s for property_id %s", report["name"], report["property_id"])
 
-                # NB: Bookmark all days with "golden" data until you find the first non-golden day
-                # - "golden" refers to data that will not change in future
-                #   requests, so we can use it as a bookmark
-                is_data_golden = raw_report_response["reports"][0]["data"].get("isDataGolden")
-                if is_historical_sync:
-                    # Switch to regular bookmarking at first golden
-                    is_historical_sync = not is_data_golden
 
-                # The assumption here is that today's data cannot be golden if yesterday's is also not golden
-                if all_data_golden and not is_historical_sync:
-                    singer.write_bookmark(state,
-                                            report["id"],
-                                            report['profile_id'],
-                                            {'last_report_date': report_date.strftime("%Y-%m-%d")})
-                    singer.write_state(state)
-                    if not is_data_golden and not is_historical_sync:
-                        # Stop bookmarking on first "isDataGolden": False
-                        all_data_golden = False
-                else:
-                    LOGGER.info("Did not detect that data was golden. Skipping writing bookmark.")
-    LOGGER.info("Done syncing %s for view_id %s", report['name'], report['profile_id'])
-
-# bookmark format
-# singer.write_bookmark(state, report["id"], report["property_id"], {"last_report_date": utils.now()})
 def sync(client, config, catalog, state):
     selected_streams = catalog.get_selected_streams(state)
-    #TODO add start with currently syncing
+    currently_syncing = state.get("currently_syncing", None)
+    selected_streams = sort_and_shuffle_streams(currently_syncing, selected_streams)
+
     for stream in selected_streams:
         state = singer.set_currently_syncing(state, stream.tap_stream_id)
         singer.write_state(state)
@@ -229,34 +210,34 @@ def sync(client, config, catalog, state):
         for field_path, field_mdata in mdata.items():
             if field_path == tuple():
                 continue
-            if field_mdata.get('inclusion') == 'unsupported':
+            if field_mdata.get("inclusion") == "unsupported":
                 continue
             _, field_name = field_path
-            if field_mdata.get('inclusion') == 'automatic' or \
-               field_mdata.get('selected') or \
-               (field_mdata.get('selected-by-default') and field_mdata.get('selected') is None):
-                if field_mdata.get('behavior') == 'METRIC':
+            if field_mdata.get("inclusion") == "automatic" or \
+               field_mdata.get("selected") or \
+               (field_mdata.get("selected-by-default") and field_mdata.get("selected") is None):
+                if field_mdata.get("behavior") == "METRIC":
                     metrics.append(Metric(name=field_name))
-                elif field_mdata.get('behavior') == 'DIMENSION':
+                elif field_mdata.get("behavior") == "DIMENSION":
                     dimensions.append(Dimension(name=field_name))
 
         end_date = get_end_date(config)
         schema = stream.schema.to_dict()
-        singer.write_schema(
-            stream.stream,
-            schema,
-            stream.key_properties
-            )
+        singer.write_schema(stream.stream,
+                            schema,
+                            stream.key_properties)
 
         report = {"property_id": config["property_id"],
+                  "account_id": config["account_id"],
                   "name": stream.stream,
                   "id": stream.tap_stream_id,
                   "metrics": metrics,
                   "dimensions": dimensions}
 
-        is_historical_sync, start_date = get_start_date(config, report['property_id'], state, report['id'])
-        sync_report(client, schema, report, start_date, end_date, state, is_historical_sync)
-        singer.write_state(state)
+        start_date = get_report_start_date(config, report["property_id"], state, report["id"])
+        request_window_size = int(config.get("request_window_size", DEFAULT_REQUEST_WINDOW_SIZE))
 
+        sync_report(client, schema, report, start_date, end_date, request_window_size, state)
+        singer.write_state(state)
     state = singer.set_currently_syncing(state, None)
     singer.write_state(state)