+++ CHANGELOG.md
@@ -1,1 +1,9 @@
 # Changelog
+## v0.0.4
+  * Update cached field exclusions [#15](https://github.com/singer-io/tap-ga4/pull/15)
+## v0.0.3
+  * Add adjustable conversion window [#13](https://github.com/singer-io/tap-ga4/pull/13)
+## v0.0.2
+  * Cache standard dimensions and metrics [#12](https://github.com/singer-io/tap-ga4/pull/12)
+## v0.0.1
+  * Alpha release [#11](https://github.com/singer-io/tap-ga4/pull/11)